moves sub-menu tests
